Creating seperated files for each the tests, + some cleaning
